make apple2 abstract class
